backend/tests/test_milestone_8.py
"""
Milestone 8: Logging, Monitoring & Analytics Tests
Comprehensive tests for all monitoring features
"""

import pytest
from fastapi.testclient import TestClient
from app.main import app
from app.core.db import SessionLocal
from app.models import User, AuditLog, LoginAttempt
from datetime import datetime, timedelta
from app.core.logging import logger, log_event, log_auth_event
from app.core.metrics import MetricsTracker
from app.services.analytics import AnalyticsService
from app.services.alerts import AlertService


client = TestClient(app)


class TestLogging:
    """Test structured logging functionality"""
    
    def test_log_event_basic(self, caplog):
        """Test basic event logging"""
        log_event(
            action="test_action",
            user_id="user123",
            target="test_target"
        )
        # Verify logging works (actual JSON output should be captured)
        assert True  # If we reach here, logging didn't crash
    
    def test_log_auth_event(self, caplog):
        """Test authentication event logging"""
        log_auth_event(
            action="login",
            user_id="user123",
            email="test@example.com",
            ip_address="127.0.0.1",
            success=True
        )
        assert True
    
    def test_log_access_event(self, caplog):
        """Test access control event logging"""
        from app.core.logging import log_access_event
        log_access_event(
            action="access_check",
            user_id="user123",
            resource="/admin/dashboard",
            allowed=False,
            details={"role": "viewer"}
        )
        assert True


class TestMetrics:
    """Test Prometheus metrics collection"""
    
    def test_track_login_attempt_success(self):
        """Test login attempt tracking - success"""
        MetricsTracker.track_login_attempt(success=True)
        # Verify metrics incremented (in real scenario, check Prometheus)
        assert True
    
    def test_track_login_attempt_failed(self):
        """Test login attempt tracking - failure"""
        MetricsTracker.track_login_attempt(success=False)
        assert True
    
    def test_track_rbac_check(self):
        """Test RBAC access check tracking"""
        MetricsTracker.track_rbac_check(allowed=True, role="admin")
        MetricsTracker.track_rbac_check(allowed=False, role="viewer")
        assert True
    
    def test_track_api_request(self):
        """Test API request tracking"""
        MetricsTracker.track_api_request(
            method="GET",
            endpoint="/users",
            status_code=200,
            duration=0.123
        )
        assert True


class TestRequestLogging:
    """Test request/response logging middleware"""
    
    def test_health_check_excluded(self):
        """Test that health checks are excluded from detailed logging"""
        response = client.get("/health")
        assert response.status_code == 200
        assert response.json() == {"status": "ok"}
    
    def test_request_id_in_state(self):
        """Test that request ID is added to request state"""
        # This would need a test endpoint that accesses request.state.request_id
        response = client.get("/health")
        assert response.status_code == 200


class TestAnalytics:
    """Test analytics service"""
    
    def test_get_users_by_role(self):
        """Test getting users by role"""
        db = SessionLocal()
        try:
            result = AnalyticsService.get_users_by_role(db)
            assert isinstance(result, dict)
        finally:
            db.close()
    
    def test_get_email_verification_stats(self):
        """Test email verification statistics"""
        db = SessionLocal()
        try:
            result = AnalyticsService.get_email_verification_stats(db)
            assert "total_users" in result
            assert "verified_users" in result
            assert "unverified_users" in result
            assert "verification_rate" in result
        finally:
            db.close()
    
    def test_get_login_stats(self):
        """Test login statistics"""
        db = SessionLocal()
        try:
            result = AnalyticsService.get_login_stats(db, hours=24)
            assert "total_attempts" in result
            assert "successful" in result
            assert "failed" in result
            assert "success_rate" in result
        finally:
            db.close()
    
    def test_get_session_stats(self):
        """Test session statistics"""
        db = SessionLocal()
        try:
            result = AnalyticsService.get_session_stats(db)
            assert "total_sessions" in result
            assert "active_sessions" in result
            assert "expired_sessions" in result
        finally:
            db.close()
    
    def test_get_audit_log_summary(self):
        """Test audit log summary"""
        db = SessionLocal()
        try:
            result = AnalyticsService.get_audit_log_summary(db, hours=24)
            assert "total_events" in result
            assert "period_hours" in result
            assert "events_by_action" in result
        finally:
            db.close()
    
    def test_get_security_dashboard(self):
        """Test comprehensive security dashboard"""
        db = SessionLocal()
        try:
            result = AnalyticsService.get_security_dashboard(db)
            assert "timestamp" in result
            assert "users" in result
            assert "email_verification" in result
            assert "login_stats" in result
            assert "session_stats" in result
        finally:
            db.close()


class TestAlerts:
    """Test alert service"""
    
    def test_check_failed_login_attempts(self):
        """Test failed login detection"""
        db = SessionLocal()
        try:
            # Create some failed login attempts
            for i in range(10):
                attempt = LoginAttempt(
                    email="attacker@example.com",
                    ip_address="192.168.1.100",
                    success=False
                )
                db.add(attempt)
            db.commit()
            
            alerts = AlertService.check_failed_login_attempts(db)
            # Should detect excessive failed attempts
            assert isinstance(alerts, list)
        finally:
            db.close()
    
    def test_check_forbidden_access_attempts(self):
        """Test forbidden access detection"""
        db = SessionLocal()
        try:
            alerts = AlertService.check_forbidden_access_attempts(db)
            assert isinstance(alerts, list)
        finally:
            db.close()
    
    def test_check_account_security(self):
        """Test account security checks"""
        db = SessionLocal()
        try:
            alerts = AlertService.check_account_security(db)
            assert isinstance(alerts, list)
        finally:
            db.close()
    
    def test_get_all_alerts(self):
        """Test getting all alerts"""
        db = SessionLocal()
        try:
            alerts = AlertService.get_all_alerts(db)
            assert isinstance(alerts, list)
        finally:
            db.close()


class TestAnalyticsEndpoints:
    """Test analytics API endpoints"""
    
    @pytest.fixture
    def admin_token(self):
        """Create admin token for testing"""
        # This would need setup of admin user and token generation
        # For now, return a placeholder
        return "test-admin-token"
    
    def test_analytics_dashboard_protected(self):
        """Test that analytics dashboard is protected"""
        response = client.get("/analytics/dashboard")
        # Should require authentication
        assert response.status_code in [401, 403]
    
    def test_analytics_alerts_protected(self):
        """Test that alerts endpoint is protected"""
        response = client.get("/analytics/alerts")
        assert response.status_code in [401, 403]


class TestMetricsEndpoint:
    """Test Prometheus metrics endpoint"""
    
    def test_metrics_endpoint_accessible(self):
        """Test that Prometheus metrics endpoint is accessible"""
        response = client.get("/metrics")
        assert response.status_code == 200
        # Should return Prometheus text format
        assert "TYPE" in response.text or "HELP" in response.text or len(response.text) > 0


# Integration tests
class TestLoggingIntegration:
    """Integration tests for logging system"""
    
    def test_request_logging_pipeline(self):
        """Test complete request logging pipeline"""
        response = client.get("/health")
        assert response.status_code == 200
        # Request should have been logged by middleware
        assert True
    
    def test_error_logging(self):
        """Test error logging"""
        # Try to access non-existent endpoint
        response = client.get("/nonexistent")
        assert response.status_code == 404
        # Error should have been logged


if __name__ == "__main__":
    pytest.main([__file__, "-v"])
